Optimized push back with move(), changed some types

+++ Heat Equationn.h
@@ -2,17 +2,19 @@
 
 #define _USE_MATH_DEFINES
 
+#include "Function.h"
+
 #include <iostream>
 #include <chrono>
 #include <fstream>
 #include <iomanip>
 #include <sstream>
-#include "Function.h"
 #include <stdexcept>
 #include <cmath>
+#include <utility>
 
 using _answer = std::pair<std::vector<std::vector<double>>, std::chrono::milliseconds>;
-using _parameteres = std::pair<int, int>;
+using _parameteres = std::pair<size_t, size_t>;
 
 // Solving Au=f for tridiagonal matrix A
 std::vector<double> SolveTridiag(std::vector<std::vector<double>>& A, std::vector<double>& f);
@@ -22,7 +24,7 @@ inline double u(double x, double t);
 
 // Comparing the real solution with approximation, return the max difference
 double Compare(std::vector<std::vector<double>> myanswer,
-	int N_x, int N_t, double thau, double h);
+	size_t N_x, size_t N_t, double thau, double h);
 
 class Solver {
 	Function f;
@@ -39,15 +41,15 @@ public:
 	Solver(const char* func, const char* d2func, double a,
 		const char* mu, const char* mu1, const char* mu2);
 
-	void addParams(int N_x, int N_t);
+	void addParams(size_t N_x, size_t N_t);
 	void solveAll();
-	void printResult(std::ostream& out, int i) const;
+	void printResult(std::ostream& out, size_t i) const;
 	int getAnswersSize() const;
 private:
 	// f + h^2/12 * f''(x, t + thau/2)
 	double half_f(double x, double t, double thau, double h) const;
 
-	void Solve(int N_x, int N_t);
+	void Solve(size_t N_x, size_t N_t);
 
 	// Initializing matrix of linear operator
 	void initA(std::vector<std::vector<double>>& A, int N_x, double thau, double h, double sigma) const;
